vikings-analytics/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cramjam==2.11.0
cryptography==46.0.4
cycler==0.12.1
fastparquet==2025.12.0
fonttools==4.61.1
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
mcp==1.26.0
nflreadpy==0.1.5
numpy==1.26.4
packaging==26.0
pandas==1.5.3
pillow==12.1.0
platformdirs==4.5.1
polars==1.38.1
polars-runtime-32==1.38.1
psycopg2-binary==2.9.11
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
